Geo/requirements.txt
# Core dependencies
asgiref==3.8.1
Django==4.2.16
django-cors-headers==3.14.0
django-extensions==3.2.3
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
django-ninja==1.3.0
geojson==3.0.0
shapely==2.0.1
redis==4.1.0

# GDAL and its dependencies
numpy>=1.0.0
wheel>=0.37.0
setuptools>=67.0.0
gdal==3.4.1

djangorestframework
psycopg2
openpyxl